+++ tests/testthat/test-package_helpers.R
@@ -11,3 +11,11 @@ test_that("handling of an unknown error works", {
     "upps"
   )
 })
+
+test_that("system information can be obtained", {
+  x <- system_information()
+  checkmate::expect_list(x)
+  checkmate::expect_names(
+    names(x), identical.to = c("machine", "cores", "ram", "os", "rversion")
+  )
+})