+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Renamed `basic_package_sticker()` -> `package_logo()` and added a demo.
 
+* Added function `matrix_diagonal_indices()` that returns the indices of the diagonal elements of a quadratic matrix.
+
 # oeli 0.4.1
 
 * Removed HMM code and `plot_sequence()` function (both do not really fit in this package).